+++ src/tests/integration/test_cases/07_invoice_rc2/03_list.test.ts
@@ -1,4 +1,5 @@
 import {
+  clearInvoiceTestContext,
   createInvoice,
   getInvoiceTestContext,
   InvoiceTestContext,
@@ -40,5 +41,9 @@ describe('Invoice RC2 - Invoice List (list both input/output invoice', () => {
     );
     expect(isOutputDataValid).toBe(true);
     expect(outputData).toBeDefined();
+  });
+
+  afterAll(async () => {
+    await clearInvoiceTestContext();
   });
 });
