+++ src/tests/integration/test_cases/07_invoice_rc2/05_output_list.test.ts
@@ -1,4 +1,5 @@
 import {
+  clearInvoiceTestContext,
   createInvoice,
   getInvoiceTestContext,
   InvoiceTestContext,
@@ -39,5 +40,9 @@ describe('Invoice RC2 - Output Invoice List', () => {
     );
     expect(isOutputDataValid).toBe(true);
     expect(outputData).toBeDefined();
+  });
+
+  afterAll(async () => {
+    await clearInvoiceTestContext();
   });
 });
